Bug 1223692: Update libvpx to v1.6.0. r=rillian
This includes the following patches:
<patch file>
<upstream review commit>
<upstream hash>

1237848-check-lookahead-ctx.patch
https://chromium-review.googlesource.com/324510
4f780e94a1fa54f22256e0f4d42a77c340a38fa1

block_error_fp.patch
https://chromium-review.googlesource.com/282611
ff8505a54d0b3dda220f5c0695519c353c82b933

cast-char-to-uint-before-shift.patch
https://chromium-review.googlesource.com/345470
2240d83d7882ce2d5d0826b9ce33b86321d7a724

clamp_abs_lvl_seg.patch
https://chromium-review.googlesource.com/315754
2e693eb80e705ea68e23eed19616d22b4778b45a

clamp-abs-QIndex.patch
https://chromium-review.googlesource.com/315802
ff3674a15e5b1a006546e1edc64c3e778eb34ab1

rename_duplicate_files.patch
https://chromium-review.googlesource.com/281967
6a82f0d7fb9ee908c389e8d55444bbaed3d54e9c
https://chromium-review.googlesource.com/317880
d36659cec7fab96cedc67db4d511ed7135637d0e

vp9_filter_restore_aligment.patch
https://chromium-review.googlesource.com/276889
33b3953c548a20c0aee705657df0440a740c28b7

vpx_once.patch
https://chromium-review.googlesource.com/312467
2635573a7f2e4bbd259379acf91efb97d983359f

MozReview-Commit-ID: R7qB6egl3Z

--HG--
rename : media/libvpx/libvpx/third_party/libyuv/source/compare_posix.cc => media/libvpx/libvpx/third_party/libyuv/source/compare_gcc.cc
rename : media/libvpx/libvpx/vp8/common/arm/neon/loopfilter_neon.c => media/libvpx/libvpx/vp8/common/arm/neon/vp8_loopfilter_neon.c
rename : media/libvpx/libvpx/vp8/common/mips/dspr2/loopfilter_filters_dspr2.c => media/libvpx/libvpx/vp8/common/mips/dspr2/vp8_loopfilter_filters_dspr2.c
rename : media/libvpx/libvpx/vp8/common/loopfilter.c => media/libvpx/libvpx/vp8/common/vp8_loopfilter.c
rename : media/libvpx/libvpx/vp8/common/x86/loopfilter_mmx.asm => media/libvpx/libvpx/vp8/common/x86/vp8_loopfilter_mmx.asm
rename : media/libvpx/libvpx/vp8/encoder/x86/quantize_sse2.c => media/libvpx/libvpx/vp8/encoder/x86/vp8_quantize_sse2.c
extra : rebase_source : 2e851aff9c848835c5598c579b9e40e0d86658d4

+++ media/libvpx/libvpx/vpx_ports/mem_ops_aligned.h
@@ -28,8 +28,8 @@
  * could redefine these macros.
  */
 #define swap_endian_16(val,raw) do {\
-    val = ((raw>>8) & 0x00ff) \
-          | ((raw<<8) & 0xff00);\
+    val = (uint16_t)(((raw>>8) & 0x00ff) \
+          | ((raw<<8) & 0xff00));\
   } while(0)
 #define swap_endian_32(val,raw) do {\
     val = ((raw>>24) & 0x000000ff) \
